backend/requirements.txt
Django==2.0.5
djangorestframework==3.8.2
mysqlclient==1.3.12
numpy==1.14.3
pandas==0.23.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
